+++ src/vs/workbench/contrib/runtimeNotebookKernel/browser/notebookExecutionStatus.ts
@@ -8,7 +8,7 @@ import { Disposable, DisposableStore, IDisposable } from '../../../../base/commo
 import { localize } from '../../../../nls.js';
 import { IConfigurationService } from '../../../../platform/configuration/common/configuration.js';
 import { IStatusbarEntryAccessor, IStatusbarService, StatusbarAlignment } from '../../../services/statusbar/browser/statusbar.js';
-import { NOTEBOOK_EXPERIMENTAL_SHOW_EXECUTION_INFO_KEY } from '../common/runtimeNotebookKernelServiceConfig.js';
+import { NOTEBOOK_EXPERIMENTAL_SHOW_EXECUTION_INFO_KEY } from '../common/runtimeNotebookKernelConfig.js';
 import { RuntimeNotebookKernel } from './runtimeNotebookKernelService.js';
 
 /**
